chore(gno.land): update content on pages (#1735)
## Description

This PR updates the content on the home page of gno.land, as well as
content in some other pages that are linked on the home page. This
change was requested by @ccomben.

The PR includes:
- Updating welcome text (intro section) on homepage
- Updating text on the "About" page
- Updating the events page - moving old to the past past events section,
adding the details of GopherCon EU
- Updating the Ecosystem page
- Reordering sections on the home page
- Updating broken and missing links
- Making the home page more intuitive to navigate

Pages have their H1 removed from the `body` because the package that
renders it (p/demo/blog) will render it twice with the current version
of the package.

This one in a series of PRs looking to improve the experience of people
visiting gno.land.

**Please, let's not bikeshed on this one** - otherwise we will get
nowhere. The Portal loop is live now, which means that a change or a fix
on any of the pages is only 1 PR away if it's needed. **Also, all
changes in this PR are minor.**

To review this PR, I highly suggest running `gnodev` and looking at the
changes rendered and comparing them to the current `gno.land` pages.

Seems running make test.sync doesn't properly update golden tests - an
issue for itself.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Morgan <git@howl.moe>