Update tutorial.md
Changed the deadline to `time.Time()` in the package and modified the deadline type in the test file. Modified conditions for deadline with `time.Now.Unix()` and added format to render, to display unix date and time correctly. I've also removed the `if success` condition in the `NewWhitelist()` function, which is not taken into account and  pass in the last `return -1` of the function even in the event of success. It's more meaningful to have a deadline in date and time, instead of in block height. In addition, the tutorial suggests visiting the `unixtimestamp` site to obtain a date in unix format.