feat(pkg/db): remove cleveldb, fsdb and rocksdb (#1714)
Split from #1602. Context:

> @zivkovicmilos:[^1] Unpopular opinion: we should probably just drop
these DBs entirely from the codebase, they are never gonna be used in
place of LevelDB in the near future
>
> @ajnavarro: IMHO the needed database is defined by the application
needs, not from any other external usecases. Sticking to one of the
existing key/value databases, we can use their special capabilities to
make it more performant for our use case (yes, even if they are all
key/value DBs, they differ in a lot of low-level functionality). \
> I would say to use one and add support in the future for more if
needed.

After a discussion in the review meeting w/ @jaekwon, this PR is now
changed to maintain another reference implementation with boltdb. It
also makes sure that it is selectable, when available, from the tm2 node
configuration.

After this PR, there are no databases in the codebase using cgo.

[^1]: https://github.com/gnolang/gno/pull/1602#discussion_r1490862277